+++ src/layers/l4-controller/commands/settings/index.ts
@@ -1,5 +1,4 @@
 export { SetUserPreferenceCommand } from './SetUserPreferenceCommand';
+export { DeleteUserPreferenceCommand } from './DeleteUserPreferenceCommand';
 export { UpdateCourseSettingsCommand } from './UpdateCourseSettingsCommand';
 export { UpdateCourseAuthorityCommand } from './UpdateCourseAuthorityCommand';
-export { SetAppSettingCommand } from './SetAppSettingCommand';
-export { DeleteAppSettingCommand } from './DeleteAppSettingCommand';